% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelDictionary.R
\name{customDict}
\alias{customDict}
\title{Generate dict of cluster names}
\usage{
customDict(sampleInfo)
}
\arguments{
\item{sampleInfo}{A data.frame contains the pixel information and cluster labels for each barcode of a human breast cancer sample.
The index contains barcodes, and at least three other columns that have these information are required and the column names should be the same as following:
"imagerow": The row pixel coordinate of the center of the spot
"imagecol": The column pixel coordinate of the center of the spot
"Cluster": The label that corresponding to this barcode}
}
\value{
A dictionary with categorical cluster labels as values
and their converted numerical labels as keys.
}
\description{
Create a dictionary with categorical cluster labels as values
and their converted numerical labels as keys
}
\examples{
fpath <- system.file("extdata", "dataframe.rda", package="stJoincount")
load(fpath)
nameList <- customDict(humanBC)

}
