% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanBC}
\alias{humanBC}
\title{Imported 10x visium public dataset - human breast cancer spatial transcriptomics sample}
\format{
A Seurat Object
}
\usage{
data("humanBC")
}
\description{
A small portion of the original human breast cancer sample is present
as the demo in this package.
}
\examples{
data("humanBC")
}
\keyword{datasets}
