% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joincount.R
\name{joincountAnalysis}
\alias{joincountAnalysis}
\title{Join count analysis.}
\usage{
joincountAnalysis(mosaicIntegration)
}
\arguments{
\item{mosaicIntegration}{RasterLayer that output from Function rasterization.}
}
\value{
Join count analysis results.
}
\description{
Join count analysis.
}
\examples{
data('humanBC')
mosaicIntegration <- rasterizeEachCluster(humanBC)
joincount.result <- joincountAnalysis(mosaicIntegration)
}
