% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterization.R
\name{mosaicIntPlot}
\alias{mosaicIntPlot}
\title{Visulization of the rasterization results}
\usage{
mosaicIntPlot(mosaicIntegration)
}
\arguments{
\item{mosaicIntegration}{Integrated mosaic of each cluster.}
}
\value{
mosaic plot with integrated pixels.
}
\description{
Visulization of the rasterization results
}
\examples{
data('humanBC')
mosaicIntegration <- rasterizeEachCluster(humanBC)
mosaicIntPlot(mosaicIntegration)

}
