% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterPrep.R
\name{rasterPrep}
\alias{rasterPrep}
\title{When sample coordinates finds a suitable buffer to ensure that each cluster is not lost
in the process of converting the spot to pixel,
apply this buffer to this function to find a suitable rectangle for the rasterlayer}
\usage{
rasterPrep(sampleInfo, n)
}
\arguments{
\item{sampleInfo}{A dataset of a human breast cancer sample containing the
pixel information and cluster labels for each barcode.}

\item{n}{buffer for extent (from function extentBuffer).}
}
\value{
raster layer with calculated resolution and extent with buffer applied
}
\description{
When sample coordinates finds a suitable buffer to ensure that each cluster is not lost
in the process of converting the spot to pixel,
apply this buffer to this function to find a suitable rectangle for the rasterlayer
}
\examples{
fpath <- system.file("script", "humanBC.rda", package="stJoincount")
load(fpath)
raster <- rasterPrep(humanBC, 15)
}
