% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{resolutionCalc}
\alias{resolutionCalc}
\title{Automatically calculating the resolution for rasterization.}
\usage{
resolutionCalc(sample)
}
\arguments{
\item{sample}{seruat object that have cluster labels attached.}
}
\value{
length and height of resolution.
}
\description{
Automatically calculating the resolution for rasterization.
}
\examples{
fpath <- system.file("extdata", "humanBC.rda", package="stJoincount")
load(fpath)
resolutionList <- resolutionCalc(humanBC)
}
