% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDataPrep.R
\name{spatialDataPrep}
\alias{spatialDataPrep}
\title{Prepare input data}
\usage{
spatialDataPrep(filename)
}
\arguments{
\item{filename}{folder pathway that contains feature-bc-mstrix, analysis and spatial information.}
}
\value{
seruat  object that have cluster labels attached.
}
\description{
Prepare input data
}
