% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscoreHeatmap.R
\name{zscorePlot}
\alias{zscorePlot}
\title{Visulization of Z-score heatmap.}
\usage{
zscorePlot(zscoreMatrix)
}
\arguments{
\item{zscoreMatrix}{calculated and reshaped z-score matirx from join count analysis.}
}
\value{
Heatmap plot
}
\description{
Visulization of Z-score heatmap.
}
\examples{
fpath <- system.file("script", "humanBC.rda", package="stJoincount")
load(fpath)
mosaicIntegration <- rasterizeEachCluster(humanBC)
joincount.result <- joincountAnalysis(mosaicIntegration)
matrix <- zscoreMatrix(humanBC, joincount.result)
zscorePlot(matrix)
}
