% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelDictionary.R
\name{customDict}
\alias{customDict}
\title{Create a dictionary with categorical cluster labels as values
and their converted numerical labels as keys}
\usage{
customDict(sampleInfo)
}
\arguments{
\item{sampleInfo}{A dataset of a human breast cancer sample containing the
pixel information and cluster labels for each barcode.}
}
\value{
A mosaic plot with labeled pixels.
}
\description{
Create a dictionary with categorical cluster labels as values
and their converted numerical labels as keys
}
\examples{
fpath <- system.file("extdata", "humanBC.rda", package="stJoincount")
load(fpath)
nameList <- customDict(humanBC)

}
