% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extentBuffer.R
\name{extentBuffer}
\alias{extentBuffer}
\title{Finding optimal number of buffer for extent.}
\usage{
extentBuffer(sample)
}
\arguments{
\item{sample}{seruat object that have cluster labels attached.}
}
\value{
optimal number of buffer for extent
}
\description{
Finding optimal number of buffer for extent.
}
\examples{
fpath <- system.file("extdata", "humanBC.rda", package="stJoincount")
load(fpath)
n <- extentBuffer(humanBC)
}
