% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterization.R
\name{rasterizeEachCluster}
\alias{rasterizeEachCluster}
\title{Converts a labeled spatial tissue map into a raster object,
in which each spatial cluster is represented by a pixel coded by label assignment.}
\usage{
rasterizeEachCluster(sampleInfo)
}
\arguments{
\item{sampleInfo}{A dataset of a human breast cancer sample containing the
pixel information and cluster labels for each barcode.}
}
\value{
A raster object converted from a labeled spatial tissue map.
}
\description{
Converts a labeled spatial tissue map into a raster object,
in which each spatial cluster is represented by a pixel coded by label assignment.
}
\examples{
fpath <- system.file("script", "humanBC.rda", package="stJoincount")
load(fpath)
mosaicIntegration <- rasterizeEachCluster(humanBC)
}
