% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{resolutionCalc}
\alias{resolutionCalc}
\title{Automatic calculation of optimal raster resolution for the sample.}
\usage{
resolutionCalc(sampleInfo)
}
\arguments{
\item{sampleInfo}{A dataset of a human breast cancer sample containing the
pixel information and cluster labels for each barcode.}
}
\value{
length and height of resolution.
}
\description{
Automatic calculation of optimal raster resolution for the sample.
}
\examples{
fpath <- system.file("extdata", "humanBC.rda", package="stJoincount")
load(fpath)
resolutionList <- resolutionCalc(humanBC)
}
