% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscoreHeatmap.R
\name{zscoreMatrix}
\alias{zscoreMatrix}
\title{This function provides a heatmap of z-scores resulting
from the join count analysis for all possible label pairs.}
\usage{
zscoreMatrix(sampleInfo, joincount.result)
}
\arguments{
\item{sampleInfo}{A dataset of a human breast cancer sample containing the
pixel information and cluster labels for each barcode.}

\item{joincount.result}{calculated result from join count analysis}
}
\value{
z-score matrix resulting from the join count analysis for all possible label pairs
}
\description{
This function provides a heatmap of z-scores resulting
from the join count analysis for all possible label pairs.
}
\examples{
fpath <- system.file("script", "humanBC.rda", package="stJoincount")
load(fpath)
mosaicIntegration <- rasterizeEachCluster(humanBC)
joincount.result <- joincountAnalysis(mosaicIntegration)
matrix <- zscoreMatrix(humanBC, joincount.result)
}
